Define opt pipeline using protobuf
This changes the opt pipeline to be built using a new OptPipelineProto. This will allow us to provide a more central place to configure passes pipelines.

The opt pipeline system is reformulated to rely on the pass-registry dependency injection system. The passes are each given a name and made available in the default registry.

This makes no changes to the opt pipeline as it currently exists. Designs will be optimized bit-identically before and after this change.

PiperOrigin-RevId: 772619437