Added a test to the IR Fuzzer that verifies the result of IR before and after an optimization pass.
A new IrFuzzTest was added which generates random IR, runs the ReassociationPass over it, plugs in parameters into the non-reassociated IR and the reassociated IR, and verifies that the results are the same. To achieve this, I added a function which accepts an optimization pass object and then proceeds to perform all of the logic. Additionally, the IrFuzzDomain was updated to allow you to generate fuzzed integer parameters, with a way to configure how many parameter sets are generated for each IR function.

PiperOrigin-RevId: 773803866